+++ src/Stryker.Core/Stryker.Core.UnitTest/MutationTest/MutationTestProcessTests.cs
@@ -1,3 +1,4 @@
+using System.CodeDom.Compiler;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.IO;
@@ -9,7 +10,6 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Moq;
 using Shouldly;
-using Stryker.Core.CoverageAnalysis;
 using Stryker.Core.Exceptions;
 using Stryker.Core.Initialisation;
 using Stryker.Core.MutantFilters;
@@ -89,24 +89,15 @@ namespace Stryker.Core.UnitTest.MutationTest
                 ExcludedMutations = new Mutator[] { }
             };
             var orchestratorMock = new Mock<BaseMutantOrchestrator<SyntaxNode>>(MockBehavior.Strict, options);
-            var reporterMock = new Mock<IReporter>(MockBehavior.Strict);
             var mutationTestExecutorMock = new Mock<IMutationTestExecutor>(MockBehavior.Strict);
-            var coverageAnalyzerMock = new Mock<ICoverageAnalyser>(MockBehavior.Strict);
 
             // setup mocks
             orchestratorMock.Setup(x => x.GetLatestMutantBatch()).Returns(mockMutants);
             orchestratorMock.Setup(x => x.Mutate(It.IsAny<SyntaxNode>())).Returns(CSharpSyntaxTree.ParseText(SourceFile).GetRoot());
             orchestratorMock.SetupAllProperties();
+            var mutator = new CsharpMutationProcess(input, fileSystem, options, null, orchestratorMock.Object);
 
-            var target = new MutationTestProcess(input,
-                reporterMock.Object,
-                mutationTestExecutorMock.Object,
-                orchestratorMock.Object,
-                fSharpOrchestrator: null,
-                fileSystem,
-                new BroadcastMutantFilter(Enumerable.Empty<IMutantFilter>()),
-                coverageAnalyzerMock.Object,
-                options);
+            var target = new MutationTestProcess(input, options, null, mutationTestExecutorMock.Object, mutator);
 
             // start mutation process
             target.Mutate();
@@ -171,13 +162,10 @@ namespace Stryker.Core.UnitTest.MutationTest
             };
 
             var orchestratorMock = new Mock<BaseMutantOrchestrator<SyntaxNode>>(MockBehavior.Strict, options);
-            var reporterMock = new Mock<IReporter>(MockBehavior.Strict);
             var mutationTestExecutorMock = new Mock<IMutationTestExecutor>(MockBehavior.Strict);
             var mutantFilterMock = new Mock<IMutantFilter>(MockBehavior.Strict);
-            var coverageAnalyzerMock = new Mock<ICoverageAnalyser>(MockBehavior.Strict);
 
             // setup mocks
-            reporterMock.Setup(x => x.OnMutantsCreated(It.IsAny<IReadOnlyProjectComponent>()));
             orchestratorMock.Setup(x => x.GetLatestMutantBatch()).Returns(mockMutants);
             orchestratorMock.Setup(x => x.Mutate(It.IsAny<SyntaxNode>())).Returns(CSharpSyntaxTree.ParseText(SourceFile).GetRoot());
             orchestratorMock.SetupAllProperties();
@@ -188,15 +176,9 @@ namespace Stryker.Core.UnitTest.MutationTest
                 .Returns((IEnumerable<Mutant> mutants, IReadOnlyFileLeaf file, StrykerOptions o) => mutants.Take(1));
 
 
-            var target = new MutationTestProcess(input,
-                reporterMock.Object,
-                mutationTestExecutorMock.Object,
-                orchestratorMock.Object,
-                fSharpOrchestrator: null,
-                fileSystem,
-                new BroadcastMutantFilter(new[] { mutantFilterMock.Object }),
-                coverageAnalyzerMock.Object,
-                options);
+            var mutator = new CsharpMutationProcess(input, fileSystem, options, new BroadcastMutantFilter(new[] { mutantFilterMock.Object }), orchestratorMock.Object);
+
+            var target = new MutationTestProcess(input, options, null, mutationTestExecutorMock.Object, mutator);
 
             // start mutation process
             target.Mutate();
@@ -253,9 +235,7 @@ namespace Stryker.Core.UnitTest.MutationTest
             // create mocks
             var options = new StrykerOptions();
             var orchestratorMock = new Mock<BaseMutantOrchestrator<SyntaxNode>>(MockBehavior.Strict, options);
-            var reporterMock = new Mock<IReporter>(MockBehavior.Strict);
             var mutationTestExecutorMock = new Mock<IMutationTestExecutor>(MockBehavior.Strict);
-            var coverageAnalyzerMock = new Mock<ICoverageAnalyser>(MockBehavior.Strict);
 
             fileSystem.AddDirectory(Path.Combine(FilesystemRoot, "TestProject", "bin", "Debug", "netcoreapp2.0"));
 
@@ -263,17 +243,10 @@ namespace Stryker.Core.UnitTest.MutationTest
             orchestratorMock.Setup(x => x.Mutate(It.IsAny<SyntaxNode>())).Returns(CSharpSyntaxTree.ParseText(SourceFile).GetRoot());
             orchestratorMock.SetupAllProperties();
             orchestratorMock.Setup(x => x.GetLatestMutantBatch()).Returns(mockMutants);
-            reporterMock.Setup(x => x.OnMutantsCreated(It.IsAny<IReadOnlyProjectComponent>()));
 
-            var target = new MutationTestProcess(input,
-                reporterMock.Object,
-                mutationTestExecutorMock.Object,
-                orchestratorMock.Object,
-                fSharpOrchestrator: null,
-                fileSystem,
-                new BroadcastMutantFilter(Enumerable.Empty<IMutantFilter>()),
-                coverageAnalyzerMock.Object,
-                options);
+            var mutator = new CsharpMutationProcess(input, fileSystem, options, null, orchestratorMock.Object);
+
+            var target = new MutationTestProcess(input, options, null, mutationTestExecutorMock.Object, mutator);
 
             target.Mutate();
 
@@ -315,12 +288,8 @@ namespace Stryker.Core.UnitTest.MutationTest
                 },
                 InitialTestRun = new InitialTestRun(scenario.GetInitialRunResult(), new TimeoutValueCalculator(500))
             };
-            var reporterMock = new Mock<IReporter>(MockBehavior.Strict);
-            reporterMock.Setup(x => x.OnMutantTested(It.IsAny<Mutant>()));
 
             var mutationExecutor = new MutationTestExecutor(scenario.GetTestRunnerMock().Object);
-
-            var mutantFilterMock = new Mock<IMutantFilter>(MockBehavior.Loose);
 
             var options = new StrykerOptions()
             {
@@ -328,11 +297,8 @@ namespace Stryker.Core.UnitTest.MutationTest
                 Concurrency = 1,
                 OptimizationMode = OptimizationModes.CoverageBasedTest
             };
-            var target = new MutationTestProcess(input,
-                reporterMock.Object,
-                mutationExecutor,
-                mutantFilter: mutantFilterMock.Object,
-                options: options);
+
+            var target = new MutationTestProcess(input, options, null, mutationExecutor);
 
             target.GetCoverage();
             target.Test(scenario.GetCoveredMutants());
@@ -381,19 +347,13 @@ namespace Stryker.Core.UnitTest.MutationTest
 
             var mutationExecutor = new MutationTestExecutor(scenario.GetTestRunnerMock().Object);
 
-            var mutantFilterMock = new Mock<IMutantFilter>(MockBehavior.Loose);
-
             var options = new StrykerOptions()
             {
                 OutputPath = basePath,
                 Concurrency = 1,
                 OptimizationMode = OptimizationModes.None
             };
-            var target = new MutationTestProcess(input,
-                reporterMock.Object,
-                mutationExecutor,
-                mutantFilter: mutantFilterMock.Object,
-                options: options);
+            var target = new MutationTestProcess(input, options, null, mutationExecutor);
 
             target.GetCoverage();
             target.Test(scenario.GetMutants());
@@ -444,8 +404,6 @@ namespace Stryker.Core.UnitTest.MutationTest
                 InitialTestRun = new InitialTestRun(scenario.GetInitialRunResult(), new TimeoutValueCalculator(500))
             };
 
-            var mutantFilterMock = new Mock<IMutantFilter>(MockBehavior.Loose);
-
             var options = new StrykerOptions
             {
                 ProjectPath = basePath,
@@ -453,11 +411,7 @@ namespace Stryker.Core.UnitTest.MutationTest
                 OptimizationMode = OptimizationModes.CoverageBasedTest
             };
 
-            var target = new MutationTestProcess(input,
-                null,
-                executor,
-                mutantFilter: mutantFilterMock.Object,
-                options: options);
+            var target = new MutationTestProcess(input, options, null, executor);
             // test mutants
             target.GetCoverage();
             
@@ -512,8 +466,6 @@ namespace Stryker.Core.UnitTest.MutationTest
                 InitialTestRun = new InitialTestRun(scenario.GetInitialRunResult(), new TimeoutValueCalculator(500))
             };
 
-            var mutantFilterMock = new Mock<IMutantFilter>(MockBehavior.Loose);
-
             var options = new StrykerOptions
             {
                 ProjectPath = basePath,
@@ -521,16 +473,13 @@ namespace Stryker.Core.UnitTest.MutationTest
                 OptimizationMode = OptimizationModes.CoverageBasedTest
             };
 
-            var target = new MutationTestProcess(input,
-                null,
-                executor,
-                mutantFilter: mutantFilterMock.Object,
-                options: options);
+            var target = new MutationTestProcess(input, options, null, executor);
+
             // test mutants
             target.GetCoverage();
             
             target.Test(input.ProjectInfo.ProjectContents.Mutants);
-            // first mutant should be marked as survided
+            // first mutant should be marked as survived
             scenario.GetMutantStatus(1).ShouldBe(MutantStatus.Survived);
         }
 
@@ -576,8 +525,6 @@ namespace Stryker.Core.UnitTest.MutationTest
                 InitialTestRun = new InitialTestRun(scenario.GetInitialRunResult(), new TimeoutValueCalculator(500))
             };
 
-            var mutantFilterMock = new Mock<IMutantFilter>(MockBehavior.Loose);
-
             var options = new StrykerOptions
             {
                 ProjectPath = basePath,
@@ -585,11 +532,7 @@ namespace Stryker.Core.UnitTest.MutationTest
                 OptimizationMode = OptimizationModes.CoverageBasedTest
             };
 
-            var target = new MutationTestProcess(input,
-                null,
-                executor,
-                mutantFilter: mutantFilterMock.Object,
-                options: options);
+            var target = new MutationTestProcess(input, options, null, executor);
             // test mutants
             target.GetCoverage();
             
@@ -638,8 +581,6 @@ namespace Stryker.Core.UnitTest.MutationTest
                 InitialTestRun = new InitialTestRun(scenario.GetInitialRunResult(), new TimeoutValueCalculator(500))
             };
 
-            var mutantFilterMock = new Mock<IMutantFilter>(MockBehavior.Loose);
-
             var options = new StrykerOptions
             {
                 ProjectPath = basePath,
@@ -647,11 +588,8 @@ namespace Stryker.Core.UnitTest.MutationTest
                 OptimizationMode = OptimizationModes.CoverageBasedTest
             };
 
-            var target = new MutationTestProcess(input,
-                null,
-                executor,
-                mutantFilter: mutantFilterMock.Object,
-                options: options);
+            var target = new MutationTestProcess(input, options, null, executor);
+
             // test mutants
             target.GetCoverage();
 
@@ -705,18 +643,12 @@ namespace Stryker.Core.UnitTest.MutationTest
                 It.IsAny<ITimeoutValueCalculator>(),
                 It.IsAny<TestUpdateHandler>()));
 
-            var mutantFilterMock = new Mock<IMutantFilter>(MockBehavior.Loose);
-
             var options = new StrykerOptions()
             {
                 ProjectPath = basePath
             };
 
-            var target = new MutationTestProcess(input,
-                reporterMock.Object,
-                executorMock.Object,
-                mutantFilter: mutantFilterMock.Object,
-                options: options);
+            var target = new MutationTestProcess(input, options, null, executorMock.Object);
 
             Should.Throw<GeneralStrykerException>(() => target.Test(input.SourceProjectInfo.ProjectContents.Mutants));
         }
@@ -747,8 +679,6 @@ namespace Stryker.Core.UnitTest.MutationTest
             executorMock.SetupGet(x => x.TestRunner).Returns(scenario.GetTestRunnerMock().Object);
             executorMock.Setup(x => x.Test(It.IsAny<IList<Mutant>>(), It.IsAny<ITimeoutValueCalculator>(), It.IsAny<TestUpdateHandler>()));
 
-            var mutantFilterMock = new Mock<IMutantFilter>(MockBehavior.Loose);
-
             var options = new StrykerOptions()
             {
                 ProjectPath = basePath
@@ -756,11 +686,7 @@ namespace Stryker.Core.UnitTest.MutationTest
 
             var reporterMock = new Mock<IReporter>(MockBehavior.Strict);
             reporterMock.Setup(x => x.OnMutantTested(It.IsAny<Mutant>()));
-            var target = new MutationTestProcess(input,
-                reporterMock.Object,
-                executorMock.Object,
-                mutantFilter: mutantFilterMock.Object,
-                options: options);
+            var target = new MutationTestProcess(input, options, null, executorMock.Object);
 
             var testResult = target.Test(input.SourceProjectInfo.ProjectContents.Mutants);
 
@@ -799,18 +725,12 @@ namespace Stryker.Core.UnitTest.MutationTest
             executorMock.SetupGet(x => x.TestRunner).Returns(runnerMock.Object);
             executorMock.Setup(x => x.Test(It.IsAny<IList<Mutant>>(), It.IsAny<ITimeoutValueCalculator>(), It.IsAny<TestUpdateHandler>()));
 
-            var mutantFilterMock = new Mock<IMutantFilter>(MockBehavior.Loose);
-
             var options = new StrykerOptions()
             {
                 ProjectPath = basePath
             };
 
-            var target = new MutationTestProcess(input,
-                reporterMock.Object,
-                executorMock.Object,
-                mutantFilter: mutantFilterMock.Object,
-                options: options);
+            var target = new MutationTestProcess(input, options, null, executorMock.Object);
 
             var testResult = target.Test(folder.Mutants);