+++ src/Stryker.Core/Stryker.Core.UnitTest/Reporters/JsonReporterTests.cs
@@ -51,7 +51,7 @@ namespace Stryker.Core.UnitTest.Reporters
         [Fact]
         public void JsonReportFileComponent_ShouldHaveLanguageSetToCs()
         {
-            var folderComponent = JsonReportTestHelper.CreateProjectWith();
+            var folderComponent = ReportTestHelper.CreateProjectWith();
             var fileComponent = (CsharpFileLeaf)(folderComponent as CsharpFolderComposite).GetAllFiles().First();
 
             new SourceFile(fileComponent).Language.ShouldBe("cs");
@@ -60,7 +60,7 @@ namespace Stryker.Core.UnitTest.Reporters
         [Fact]
         public void JsonReportFileComponent_ShouldContainOriginalSource()
         {
-            var folderComponent = JsonReportTestHelper.CreateProjectWith();
+            var folderComponent = ReportTestHelper.CreateProjectWith();
             var fileComponent = (CsharpFileLeaf)(folderComponent as CsharpFolderComposite).GetAllFiles().First();
 
             new SourceFile(fileComponent).Source.ShouldBe(fileComponent.SourceCode);
@@ -69,7 +69,7 @@ namespace Stryker.Core.UnitTest.Reporters
         [Fact]
         public void JsonReportFileComponents_ShouldContainMutants()
         {
-            var folderComponent = JsonReportTestHelper.CreateProjectWith();
+            var folderComponent = ReportTestHelper.CreateProjectWith();
             foreach (var file in (folderComponent as CsharpFolderComposite).GetAllFiles())
             {
                 var jsonReportComponent = new SourceFile(((CsharpFileLeaf)file));
@@ -84,7 +84,7 @@ namespace Stryker.Core.UnitTest.Reporters
         public void JsonReportFileComponent_DoesNotContainDuplicateMutants()
         {
             var loggerMock = Mock.Of<ILogger>();
-            var folderComponent = JsonReportTestHelper.CreateProjectWith(duplicateMutant: true);
+            var folderComponent = ReportTestHelper.CreateProjectWith(duplicateMutant: true);
             foreach (var file in (folderComponent as CsharpFolderComposite).GetAllFiles())
             {
                 var jsonReportComponent = new SourceFile(((CsharpFileLeaf)file), loggerMock);
@@ -98,7 +98,7 @@ namespace Stryker.Core.UnitTest.Reporters
         [Fact]
         public void JsonReport_ThresholdsAreSet()
         {
-            var folderComponent = JsonReportTestHelper.CreateProjectWith();
+            var folderComponent = ReportTestHelper.CreateProjectWith();
 
             var report = JsonReport.Build(new StrykerOptions(), folderComponent, It.IsAny<TestProjectsInfo>());
 
@@ -110,7 +110,7 @@ namespace Stryker.Core.UnitTest.Reporters
         [Fact]
         public void JsonReport_ShouldContainAtLeastOneFile()
         {
-            var folderComponent = JsonReportTestHelper.CreateProjectWith();
+            var folderComponent = ReportTestHelper.CreateProjectWith();
 
             var report = JsonReport.Build(new StrykerOptions(), folderComponent, It.IsAny<TestProjectsInfo>());
 
@@ -120,7 +120,7 @@ namespace Stryker.Core.UnitTest.Reporters
         [Fact]
         public void JsonReport_ShouldContainTheProjectRoot()
         {
-            var folderComponent = JsonReportTestHelper.CreateProjectWith();
+            var folderComponent = ReportTestHelper.CreateProjectWith();
 
             var report = JsonReport.Build(new StrykerOptions(), folderComponent, It.IsAny<TestProjectsInfo>());
 
@@ -139,7 +139,7 @@ namespace Stryker.Core.UnitTest.Reporters
             };
             var reporter = new JsonReporter(options, mockFileSystem);
 
-            reporter.OnAllMutantsTested(JsonReportTestHelper.CreateProjectWith(), It.IsAny<TestProjectsInfo>());
+            reporter.OnAllMutantsTested(ReportTestHelper.CreateProjectWith(), It.IsAny<TestProjectsInfo>());
             var reportPath = Path.Combine(options.ReportPath, "mutation-report.json");
             mockFileSystem.FileExists(reportPath).ShouldBeTrue($"Path {reportPath} should exist but it does not.");
             var fileContents = mockFileSystem.File.ReadAllText(reportPath);