+++ src/Stryker.Core/Stryker.Core/MutationTest/FsharpMutationProcess.cs
@@ -2,11 +2,9 @@ using Microsoft.Extensions.Logging;
 using Microsoft.FSharp.Collections;
 using Stryker.Core.Compiling;
 using Stryker.Core.Logging;
-using Stryker.Core.MutantFilters;
 using Stryker.Core.Mutants;
 using Stryker.Core.Options;
 using Stryker.Core.ProjectComponents;
-using Stryker.Core.Reporters;
 using System;
 using System.IO;
 using System.IO.Abstractions;
@@ -15,7 +13,7 @@ using static FSharp.Compiler.SyntaxTree;
 
 namespace Stryker.Core.MutationTest
 {
-    class FsharpMutationProcess : IMutationProcess
+    public class FsharpMutationProcess : IMutationProcess
     {
         private readonly ProjectComponent<ParsedInput> _projectInfo;
         private readonly ILogger _logger;
@@ -23,12 +21,19 @@ namespace Stryker.Core.MutationTest
         private readonly FsharpCompilingProcess _compilingProcess;
         private readonly BaseMutantOrchestrator<FSharpList<SynModuleOrNamespace>> _orchestrator;
 
+        /// <summary>
+        /// This constructor is for tests
+        /// </summary>
+        /// <param name="mutationTestInput"></param>
+        /// <param name="fileSystem"></param>
+        /// <param name="options"></param>
+        /// <param name="orchestrator"></param>
         public FsharpMutationProcess(MutationTestInput mutationTestInput,
-            BaseMutantOrchestrator<FSharpList<SynModuleOrNamespace>> orchestrator = null,
-            IFileSystem fileSystem = null,
-            StrykerOptions options = null)
+            IFileSystem fileSystem,
+            StrykerOptions options,
+            BaseMutantOrchestrator<FSharpList<SynModuleOrNamespace>> orchestrator)
         {
-            MutationTestInput input = mutationTestInput;
+            var input = mutationTestInput;
             _projectInfo = (ProjectComponent<ParsedInput>)input.SourceProjectInfo.ProjectContents;
 
             _options = options;
@@ -37,6 +42,14 @@ namespace Stryker.Core.MutationTest
             _logger = ApplicationLogging.LoggerFactory.CreateLogger<MutationTestProcess>();
         }
 
+        /// <summary>
+        /// This constructor is used by the <see cref="MutationTestProcess"/> initialization logic.
+        /// </summary>
+        /// <param name="mutationTestInput"></param>
+        /// <param name="options"></param>
+        public FsharpMutationProcess(MutationTestInput mutationTestInput,
+            StrykerOptions options): this(mutationTestInput, null, options, null){}
+
         public void Mutate()
         {
             // Mutate source files
@@ -44,15 +57,15 @@ namespace Stryker.Core.MutationTest
             {
                 _logger.LogDebug($"Mutating {file.RelativePath}");
                 // Mutate the syntax tree
-                var treeroot = ((ParsedInput.ImplFile)file.SyntaxTree).Item.modules;
-                var mutatedSyntaxTree = _orchestrator.Mutate(treeroot);
+                var treeRoot = ((ParsedInput.ImplFile)file.SyntaxTree).Item.modules;
+                var mutatedSyntaxTree = _orchestrator.Mutate(treeRoot);
                 // Add the mutated syntax tree for compilation
                 var tree = (ParsedInput.ImplFile)file.SyntaxTree;
                 var item = tree.Item;
                 //we hardcode the lastcompiled flag to make the compile pass
-                //this needs to be fixed in the FSharp.Compiler.SourceCodeServices package, or made dynamic as it now assumes the bottom of Program.fs is the entrypoint
-                var lastcompiled = item.fileName.Equals("Program.fs") ? new Tuple<bool, bool>(true, true) : item.isLastCompiland;
-                file.MutatedSyntaxTree = ParsedInput.NewImplFile(ParsedImplFileInput.NewParsedImplFileInput(item.fileName, item.isScript, item.qualifiedNameOfFile, item.scopedPragmas, item.hashDirectives, mutatedSyntaxTree, lastcompiled));
+                //this needs to be fixed in the FSharp.Compiler.SourceCodeServices package, or made dynamic as it now assumes the bottom of Program.fs is the entry point
+                var lastCompile = item.fileName.Equals("Program.fs") ? new Tuple<bool, bool>(true, true) : item.isLastCompiland;
+                file.MutatedSyntaxTree = ParsedInput.NewImplFile(ParsedImplFileInput.NewParsedImplFileInput(item.fileName, item.isScript, item.qualifiedNameOfFile, item.scopedPragmas, item.hashDirectives, mutatedSyntaxTree, lastCompile));
                 if (_options.DevMode)
                 {
                     _logger.LogTrace($"Mutated {file.RelativePath}:{Environment.NewLine}{mutatedSyntaxTree}");
@@ -94,8 +107,6 @@ namespace Stryker.Core.MutationTest
 
         public void FilterMutants()
         {
-            //fsharp implementation is work in progress
-            throw new NotImplementedException();
         }
     }
 }