+++ src/Stryker.Core/Stryker.Core/TestRunners/VsTest/VsTestContextInformation.cs
@@ -120,6 +120,7 @@ namespace Stryker.Core.TestRunners.VsTest
             }
             catch (Exception e)
             {
+                _logger.LogError("Stryker failed to connect to vstest.console with error: {error}", e.Message);
                 throw new GeneralStrykerException("Stryker failed to connect to vstest.console", e);
             }
             return vsTestConsole;